Some minor fixes:  - Fixed Events descritpion  - shifted some classes out of main.dart for readability  - put a default thumbnail & post cover
