+++ lib/main.dart
@@ -6,6 +6,9 @@ import 'package:http/http.dart' as http;
 import 'package:share/share.dart';
 import 'package:html_unescape/html_unescape.dart';
 
+import 'CustomWidgets.dart';
+import 'Helpers.dart';
+
 void main() => runApp(new MyApp());
 
 //Starting Activity
@@ -87,26 +90,6 @@ class FLHome extends StatelessWidget {
       },
     );
 
-    // builder for buttons
-    Column buildButtonColumn(IconData icon, String label) {
-      final color = Colors.deepOrange;
-      return new Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          new Icon(icon, color: color),
-          new Container(
-              margin: const EdgeInsets.only(top: 8.0),
-              child: new Text(
-                label,
-                style: new TextStyle(
-                  fontSize: 12.0,
-                  fontWeight: FontWeight.w400,
-                  color: color,
-                ),
-              )),
-        ],
-      );
-    }
 
     //Row with buttons
     Widget buttons = new Container(
@@ -115,25 +98,25 @@ class FLHome extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         children: <Widget>[
           new GestureDetector(
-            child: buildButtonColumn(Icons.phone, 'Chiama'),
+            child: new CustomButton(Icons.phone, 'Chiama'),
             onTap: () async {
               await launch("tel://0437851357");
             },
           ),
           new GestureDetector(
-            child: buildButtonColumn(Icons.mail, 'Scrivici'),
+            child: new CustomButton(Icons.mail, 'Scrivici'),
             onTap: () async {
               await launch("mailto:fablab@centroconsorzi.it");
             },
           ),
           new GestureDetector(
-            child: buildButtonColumn(Icons.directions, 'Indicazioni'),
+            child: new CustomButton(Icons.directions, 'Indicazioni'),
             onTap: () async {
               await launch("geo:0,0?q=Fablab Belluno");
             },
           ),
           new GestureDetector(
-            child: buildButtonColumn(Icons.web, 'Sito Web'),
+            child: new CustomButton(Icons.web, 'Sito Web'),
             onTap: () async {
               await launch("http://fablab.centroconsorzi.it/");
             },
@@ -223,17 +206,7 @@ class PostItem extends StatelessWidget {
       return new ExpansionTile(
           title: new Row(
             children: <Widget>[
-              post.thumbnailUrl != null
-                  ? // check if there is a image
-                  new ClipOval(
-                      //creating oval cliph for the thumbnail
-                      child: new Image.network(
-                          //retrieving the image from the network
-                          post.thumbnailUrl,
-                          height: 50.0,
-                          fit: BoxFit.fill),
-                    )
-                  : new Column(),
+              new ThumbnailImage(post.thumbnailUrl),
               new Expanded(
                   child: new Container(
                       margin: new EdgeInsets.only(left: 20.0),
@@ -318,108 +291,31 @@ class FLResponse {
   }
 }
 
-class Post {
-  final int id;
-  final String url;
-  final String title;
-  final String content;
-  final String thumbnailUrl;
-  final String medThumb;
-  final String publishDate;
-  final String author;
 
-  Post(
-      {this.id,
-      this.url,
-      this.title,
-      this.content,
-      this.thumbnailUrl,
-      this.medThumb,
-      this.publishDate,
-      this.author});
-
-  factory Post.fromJson(Map<String, dynamic> json) {
-    String thumb = json['thumbnail'];
-    if (thumb != null) {
-      final thumbs = json['thumbnail_images'];
-      thumb = thumbs['medium_large']['url'];
-    }
-
-    String excerpt = json['excerpt'];
-    excerpt = excerpt.substring(3);
-    excerpt = excerpt.substring(0, excerpt.length - 5);
-
-    String pDate = json['modified'];
-    pDate = pDate.substring(0, 10);
-    pDate = pDate.split('-').reversed.join('/');
-
-    String authorName = json['author']['first_name'] != ""
-        ? json['author']['first_name'] + " " + json['author']['last_name']
-        : json['author']['name'];
-
-    return new Post(
-        title: json['title'],
-        content: excerpt,
-        url: json['url'],
-        id: json['id'],
-        thumbnailUrl: json['thumbnail'],
-        medThumb: thumb,
-        publishDate: pDate,
-        author: authorName);
-  }
-}
 
 //SINGLE article page
 
 class SingleArticle extends StatelessWidget {
-  Post post;
-  SingleArticle(Post post) {
-    this.post = post;
-  }
+  final Post post;
+
+  const SingleArticle(this.post); //creating from poassing a Post object
+
+
 
   @override
   Widget build(BuildContext context) {
-    Column buildButtonColumn(IconData icon, String label) {
-      final color = Colors.deepOrange;
-      return new Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          new Icon(icon, color: color),
-          new Container(
-              margin: const EdgeInsets.only(top: 8.0),
-              child: new Text(
-                label,
-                style: new TextStyle(
-                  fontSize: 12.0,
-                  fontWeight: FontWeight.w400,
-                  color: color,
-                ),
-              )),
-        ],
-      );
-    }
-
-    Widget cover = post.thumbnailUrl != null
-        ? new Image.network(
-            post.medThumb,
-            width: MediaQuery.of(context).size.width,
-            fit: BoxFit.cover,
-            alignment: Alignment.topCenter,
-          )
-        : new Column();
-
     Row buttons = new Row(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         crossAxisAlignment: CrossAxisAlignment.end,
         children: <Widget>[
           new GestureDetector(
-            child: buildButtonColumn(Icons.web, 'Leggi Tutto'),
+            child: new CustomButton(Icons.web, 'Leggi Tutto'),
             onTap: () async {
               await launch(post.url);
             },
           ),
           new GestureDetector(
-            child: buildButtonColumn(Icons.share, 'Condividi'),
+            child: new CustomButton(Icons.share, 'Condividi'),
             onTap: () async {
               await share(post.url);
             },
@@ -436,7 +332,8 @@ class SingleArticle extends StatelessWidget {
         child: new Column(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
-            new ClipPath(child: cover, clipper: new BottomWaveClipper()),
+            new ClipPath(child: new CoverImage(post.medThumb),
+                clipper: new BottomWaveClipper()),
             new Container(
               child: new Text(
                 post.title,
@@ -463,10 +360,11 @@ class SingleArticle extends StatelessWidget {
                   ),
                   new Expanded(
                       child: new Text(
-                    post.publishDate,
-                    textAlign: TextAlign.end,
-                    style: new TextStyle(color: Colors.grey[700]),
-                  )),
+                        post.publishDate,
+                        textAlign: TextAlign.end,
+                        style: new TextStyle(color: Colors.grey[700]),
+                      )
+                  ),
                 ],
               ),
             ),
@@ -503,6 +401,7 @@ class BottomWaveClipper extends CustomClipper<Path> {
 
 //articles helper ends ==============================================
 
+
 class Events extends StatelessWidget {
   final uri = 'http://fablab.centroconsorzi.it/api/';
 
@@ -538,63 +437,7 @@ class Events extends StatelessWidget {
 
 //------------------------------- events getter
 
-class EventPost extends Post {
-  final int id;
-  final String url;
-  final String title;
-  final String content;
-  final String thumbnailUrl;
-  final String time;
-  final String geo;
-  final String medThumb;
 
-  EventPost({
-    this.id,
-    this.url,
-    this.title,
-    this.content,
-    this.thumbnailUrl,
-    this.time,
-    this.geo,
-    this.medThumb,
-  });
-
-  factory EventPost.fromJson(Map<String, dynamic> json) {
-    final dateList = json['custom_fields']['data'];
-    final coords = json['custom_fields']['luogo'];
-
-    String time = '';
-    if (dateList != null) {
-      time = dateList[0];
-      String yyyy = time.substring(0, 4);
-      String mm = time.substring(4, 6);
-      String dd = time.substring(6);
-      time = dd + ' - ' + mm + ' - ' + yyyy;
-    }
-    String tempgc = '';
-    if (coords != null) {
-      tempgc = coords[0];
-    }
-    String thumb = json['thumbnail'];
-    if (thumb != null) {
-      final thumbs = json['thumbnail_images'];
-      thumb = thumbs['medium_large']['url'];
-    }
-    String excerpt = json['excerpt'];
-    excerpt = excerpt.substring(3);
-    excerpt = excerpt.substring(0, excerpt.length - 5);
-    return new EventPost(
-      time: time,
-      title: json['title'],
-      content: excerpt,
-      url: json['url'],
-      id: json['id'],
-      thumbnailUrl: json['thumbnail'],
-      medThumb: thumb,
-      geo: tempgc,
-    );
-  }
-}
 
 class EventPostItem extends StatelessWidget {
   const EventPostItem(this.post);
@@ -606,11 +449,7 @@ class EventPostItem extends StatelessWidget {
       return new ExpansionTile(
           title: new Row(
             children: <Widget>[
-              post.thumbnailUrl != null
-                  ? new ClipOval(
-                      child: new Image.network(post.thumbnailUrl,
-                          height: 50.0, fit: BoxFit.cover))
-                  : new Column(),
+              new ThumbnailImage(post.thumbnailUrl),
               new Expanded(
                   child: new Container(
                 margin: new EdgeInsets.only(left: 20.0),
@@ -664,7 +503,6 @@ class EventPostItem extends StatelessWidget {
             )
           ]);
     }
-
     return _buildTiles(post);
   }
 }
