+++ pubspec.yaml
@@ -31,6 +31,7 @@ flutter:
   uses-material-design: true
   assets:
       - images/fablogo.bmp
+      - images/fablogo.png
 
   # To add assets to your application, add an assets section, like this:
   # assets:
